Add JWTKeyCommand and update JWTServiceProvider with command registration
This commit introduces a new console command, JWTKeyCommand, in the SimonMarcelLinden\JWT\Console\Commands namespace.
The command, defined with the signature 'jwt:generate', is designed to generate and set a new JWT secret key in the .env file.
In addition to this, the commit includes updates to JWTServiceProvider, registering the new JWTKeyCommand for console execution.
It also involves the removal of .gitkeep files in various directories (Auth/Providers, Console/Commands, Http/Middleware)
as these directories now contain implemented classes.