Update JWTGuard and JwtUserProvider for LDAP Authentication
This commit includes updates to the config, JWTGuard and JwtUserProvider classes to add LDAP authentication support.
LDAP authentication can be enabled via configuration, and the JWTGuard now checks LDAP credentials if the user is not found in the database. The JwtUserProvider also includes a new method to create a user from LDAP credentials.

- src/Auth/Guards/JWTGuard.php
- src/Auth/Providers/JwtUserProvider.php
- config/config.php

---

This commit includes updates to the JWTGuard and JwtUserProvider classes to support LDAP authentication as an alternative authentication method. Additionally, the configuration file "config.php" has been modified to allow specifying the authentication method and LDAP settings.